trackPrecompute 6

+++ include/viewer.h
@@ -916,7 +916,7 @@ private:
     size_t m_patchSize{0};
     size_t m_nPoints{0};
     size_t m_patchArea{0};
-    size_t m_patchCenter{0};
+    float m_patchCenter{0.0f};
 
     float m_scaleFactor{1.0f};
     std::vector<float> m_invScaleFactors;
